Move HomePage to screens folder
Refs #380.
